It's code for secret auction programm.
Its very useful while an auction to hide the bidders and the bidded amount from upcoming bidders. So that they will try to bid higher to get the perticular owned. And keeping the logic in mind I made a python project to fulfill the requirement. I hope it will for sure help.